test(incremental): split incremental test into separate file w/ own DB
This isolates the incremental test from being affected by the side
effects from the e2e tests.
